lib: Pass down boot vector for secondary CPUs as capabilities
Pass the boot vector and boot data for secondary processors
around as capabilties not plain addresses.

+++ include/sbi/sbi_hsm.h
@@ -63,7 +63,7 @@ void __noreturn sbi_hsm_exit(struct sbi_scratch *scratch);
 
 int sbi_hsm_hart_start(struct sbi_scratch *scratch,
 		       const struct sbi_domain *dom,
-		       u32 hartid, ulong saddr, ulong smode, ulong arg1);
+		       u32 hartid, uintptr_t saddr, ulong smode, uintptr_t arg1);
 int sbi_hsm_hart_stop(struct sbi_scratch *scratch, bool exitnow);
 void sbi_hsm_hart_resume_start(struct sbi_scratch *scratch);
 void __noreturn sbi_hsm_hart_resume_finish(struct sbi_scratch *scratch,
